move `to_scancode` and `from_scancode` builtin KeyCode
