+++ src/keyboard.rs
@@ -73,6 +73,7 @@ use nameof::name_of;
 use std::{fmt, str::FromStr};
 
 use crate::error::OsError;
+use crate::platform_impl::{keycode_from_scancode, keycode_to_scancode};
 
 impl ModifiersState {
   /// Returns `true` if the shift key is pressed.
@@ -672,6 +673,15 @@ pub enum KeyCode {
   F34,
   /// General-purpose function key.
   F35,
+}
+
+impl KeyCode {
+  pub fn to_scancode(self) -> Option<u32> {
+    keycode_to_scancode(self)
+  }
+  pub fn from_scancode(scancode: u32) -> KeyCode {
+    keycode_from_scancode(scancode)
+  }
 }
 
 impl FromStr for KeyCode {
